Adding a local vault check
index.html now includes a static js file from localhost. If it loads correctly, and the current url is not localhost or 127.0.0.1, a link for "Local vault" shows up in the toolbar. If not, it shows nothing.

The downside of the current system is that, anytime you use vault remotely when you do not have a local vault, the file won't load, and you'll get an error (or warning symbol, or something) in the browser's status bar.